+++ circus-train-s3-s3-copier/src/test/java/com/hotels/bdp/circustrain/s3s3copier/S3S3CopierOptionsTest.java
@@ -96,4 +96,24 @@ public class S3S3CopierOptionsTest {
     assertNull(options.getCannedAcl());
   }
 
+  @Test
+  public void getMaxCopyAttempts() throws Exception {
+    copierOptions.put(S3S3CopierOptions.Keys.MAX_COPY_ATTEMPTS.keyName(), 3);
+    S3S3CopierOptions options = new S3S3CopierOptions(copierOptions);
+    assertThat(options.getMaxCopyAttempts(), is(3));
+  }
+
+  @Test
+  public void getMaxCopyAttemptsDefaultIsThree() throws Exception {
+    S3S3CopierOptions options = new S3S3CopierOptions(copierOptions);
+    assertThat(options.getMaxCopyAttempts(), is(3));
+  }
+
+  @Test
+  public void getMaxCopyAttemptsDefaultIfLessThanOne() throws Exception {
+    copierOptions.put(S3S3CopierOptions.Keys.MAX_COPY_ATTEMPTS.keyName(), -1);
+    S3S3CopierOptions options = new S3S3CopierOptions(copierOptions);
+    assertThat(options.getMaxCopyAttempts(), is(3));
+  }
+
 }